devbox: clean up profile history after sync
After syncing the flake packages to the profile, remove all old
generations of the Nix profile. This allows the Nix garbage collector to
eventually remove any old packages.

Opted for a Go implementation instead of calling `nix profile
wipe-history` because deleting the history is pretty simple and this is
faster.