patchpkg: fix system CUDA lib search (#2507)
As part of the auto-patching for CUDA, devbox searches for the system's
libcuda (installed by the driver) and adds it to the Nix store. This
fixes a couple of bugs with that search process:

- When creating the soname links, the base name of the path to libcuda
wasn't being used. This would lead to creating symlinks like
"lib/libcuda.so.1", which would fail because a lib subdirectory didn't
exist.
- Ensure the `src` attribute is set in the patch flake so that devbox
knows the path the flake source (which contains the copied libcuda) at
build time.