all: upgrade to Go 1.24
Upgrade to Go 1.24 and use `go get -tool` to manage golangci-lint and
gofumpt instead of runx. They're installed to `dist/tools` (which is
also prepended to PATH) when a devbox shell starts.

Added `git` and `fd` to `devbox.json` since we have scripts that rely on
them.

Other misc. changes fix the warnings from the latest version of
golangci-lint.