[flake] fix vendor hash (#2548)
## Summary

## How was it tested?

## Community Contribution License

All community contributions in this pull request are licensed to the
project
maintainers under the terms of the
[Apache 2 License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request, I represent that I have the right to
license the
contributions to the project maintainers under the Apache 2 License as
stated in
the
[Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Chores**
- Updated the value used to verify the integrity of external
dependencies. This update ensures that underlying resources are based on
the most recent information, contributing to improved system security
and consistency. The changes do not affect any user-facing functionality
and are part of routine maintenance to guarantee updated assets.
Overall, this improvement bolsters system reliability.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->