Update poetry demo due to major version change
Poetry is now on major version 2 and has had some significant changes.

The `poetry shell` command is deprecated and `[tool.poetry.scripts]` no
longer are available on the path.

This commit updates updates the docs and the initHook to ensure that
scripts are available.

It makes the compromise of using poetry to manage the installed python
version instead of devbox

Issue: [Poetry shell missing scripts](https://github.com/jetify-com/devbox/issues/2676)