Remove docs and documentation app (#2724)
## Summary
Deleted all files related to documentation, including configuration,
markdown, images, and source files under the docs/ and docs/app/
directories. This removes the Docusaurus-based documentation site and
its supporting assets from the repository.

## How was it tested?

## Community Contribution License

All community contributions in this pull request are licensed to the
project
maintainers under the terms of the
[Apache 2 License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request, I represent that I have the right to
license the
contributions to the project maintainers under the Apache 2 License as
stated in
the
[Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).