Fix broken link for 'Installing Devbox' (#2730)
## Summary
`Install Devbox` in multiple README files points to
https://www.jetify.com/docs/devbox/installing_devbox which throws a 404
error.
This PR fix the broken links and updates with the new ones.

## How was it tested?
Looking and following the links that were broken

## Community Contribution License

All community contributions in this pull request are licensed to the
project
maintainers under the terms of the
[Apache 2 License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request, I represent that I have the right to
license the
contributions to the project maintainers under the Apache 2 License as
stated in
the
[Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).

Signed-off-by: Abolfazl Shahbazi <12436063+ashahba@users.noreply.github.com>