Fix VS Code "Reopen in Devbox shell environment" in WSL (#2729)
## Summary

Fixes #1319 - the "Reopen in Devbox shell environment" command does not
work in a VS Code WSL window

## How was it tested?

Tested on Windows (WSL), Linux and MacOS. In Windows, the WSL VS Code
extension must be installed, and the test project must reside in WSL.

1. Build Devbox and place it in the `PATH`
2. Run VS Code
3. Open a project that has a `devbox.json` file at the root
4. In a Terminal window, verify that the packages in `devbox.json` are
not executable (not in the `PATH`)
5. Open the command palette and select "Devbox: Reopen in Devbox shell
environment"
6. Verify that the VS Code window closes, and a new VS Code window is
opened for the same project.
7. In a Terminal window, exit out of any Devbox shells and verify that
the packages in `devbox.json` are executable.

## Community Contribution License

All community contributions in this pull request are licensed to the
project
maintainers under the terms of the
[Apache 2 License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request, I represent that I have the right to
license the
contributions to the project maintainers under the Apache 2 License as
stated in
the
[Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).